+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Svizzle changelog
 
+## next
+
+## `@svizzle/barchart` v0.11.0 (next)
+
+- changed the shape of the `theme` prop
+	- removed `theme.fontSize`
+	- added `theme.glyphHeight` and `theme.glyphWidth`
+
+
 ## 20230907
 
 ## `@svizzle/histogram` v0.6.5